requirements.txt
#Requiremnets - yes i could lock dwon wih versions but Streamlits Package manger will resolve conflicts

streamlit== 1.32.0  
pandas==2.2.2
numpy==1.26.4
matplotlib.pyplot
hmac
pystac_client==0.8.1 
odc.stac==0.3.9  

